+++ src/LumengPHP/Db/Connection/PDOProviderInterface.php
@@ -5,11 +5,11 @@ namespace LumengPHP\Db\Connection;
 use PDO;
 
 /**
- * PDO工厂接口
+ * PDO Provider 接口
  *
  * @author zhengluming <luming.zheng@shandjj.com>
  */
-interface PDOFactoryInterface {
+interface PDOProviderInterface {
 
     /**
      * 服务器类型：主服务器
@@ -27,5 +27,5 @@ interface PDOFactoryInterface {
      * @param int $type 服务器类型
      * @return PDO
      */
-    public function getPDO($type = PDOFactoryInterface::TYPE_MASTER);
+    public function getPDO($type = PDOProviderInterface::TYPE_MASTER);
 }